requirements.txt
Django==4.2.4
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2==2.9.7
gunicorn==20.1.0
django-cors-headers==4.3.1
Pillow==10.1.0